Initial Commit
The Udemy aspnet-core-angular class is based on 1.1. Microsoft is close to pushing out 2.0. This project is set to SDK 1.0.4 and runtime 1.1.2 (hopefully). I made a 2.0 project, due to the fact that MS is pushing 2.0 EF libs by default, but I was having problems with migrations using EF tools - it would not pick up my connection strings from appsettings.json. So I am going to try with 1.1 again and go through the pain of specifying version numbers.