Added Angular UI for New Vehicles
Started the Angular forms to make this app. So this app is using Angular 4 which did away with relative urls for endpoint calls. I have yet to find a good solution online. Docs and discussions seem to be all weaksauce. It's insane that this is as popular as it is. I wonder if React is better.