Refactored VehicleResource to Update and Query Types
One resource for both query and update gets the job done but it is not efficient. An update resource should have the minimum required to update records, i.e. keys and foreign keys. A query resource should not include foreign keys or any other unnecessary internal data. So VehicleResource was renamed to SaveVehicleResource and a new VehicleResource was added to return query data. The necessary mappings and eager loading in the VehiclesController are added.