Set Up Toasts for Errors
This was a battle! The ng2-toasty used in the class video did not seem to work for Angular 4. No errors would display, no exceptions thrown, despite the toast code being executed. Ngx-toastr was a bit of a struggle and requires a lot more wiring, but it works! I'll try ng2-toastr next, as it seems to also have Angular 4 support and looks simpler to implement.