Fixed Merge Changes Again for Ng2-Toastr
I wanted a clean copy with ng2-toastr working. Due to some cyclic dependencies in the app.module files alleged by Angular when trying to load the ToastsManager in a global custom error handler that I am unable to fix, I'm going to try to revert back to ngx-toastr after this.