Set Global Error Handler to Display Toasts
It works! Needs to use an ngZone to wrap the ToastyService call so that the toast will show up properly; ngZone is a call to setTimeout with an after effect of triggering Angular's change detection. Wrapped that in an if statement that checks for a real window object in order to get past server-side prerendering.