Added Vehicle View UI
Moved the existing vehicle view out of the new vehicle form and into it's own component for read-only viewing. Moved the delete to that component. Still using the new vehicle form to actual data edits. The view component includes two tabs, basic (for the existing data we have), and photos for the new photo upload feature we will implement.