Started Implementing Photo Uploads
Added domain entities for uploading photos. Keeping a GUID file name for the images with the provided extension. Will add validation for valid file extensions. Implemented backend endpoint for uploads. Need to make one for file image URL. The files are stored in the file system - wwwroot/uploads. The DB just keeps a file name and the app assumes the files are all located in the same folder in the file system.