Added Photo Upload to the UI and Fixed Tab Bug
Wired up the form and services to upload photos to the new endpoint.
Added an import to boot-client.js to fix a bug where clicking on any tab (basic or photos) would trigger the app to route, which would fail and show the default. Importing "bootstrap" made the tabs work correctly, no page reloads.