Update Domain Model and Added More Client-Side Logging
Added an explicit foreign key to the Photo entity back to Vehicle.
Updated the UI to consume the GetPhotos endpoint.
Added dependency injection for the photo repo. The startup is a little buggy, might need to restart it again if it fails to link the interface with the implementation.
Added remote logging to a resource loading code.