Implemented Progress Bar
O. M. G. This was a crazy battle. But I learned a lot about the dependency injection in Angular. I was getting issues before where I could not call subcribe to any methods on services passed into components via their constructor. A solution I had found was to put in a providers property for the @Component with the service. And that worked! But with the ProgressService it was just not working. Turns out that doing this makes another copy of the service per component, and for the ProgressService I needed the one copy in the app module to be injected instead. So I got rid of that provider in the @Component. It worked this time, but I'll have to deal with it if it happens again.