Secured App But Broken ProgressService
Finished securing the app. Only authenticated users can make edits, including uploading photos.
The AuthHttp is not compatible with ProgressService. As a work-around, Mosh had us move the provider from the app module to the vehicle-view component. However, when added that way, it does not override the default BrowserHrx used by the photo service; i.e. it looks like it's being called, but the progress events never tick because the default BrowserHrx is being used, not the custom one. Idk how to fix that yet, nor do I have time right now.